[KOGITO-7461] Refactoring events api
KogitoEventExecutor does not belong there
