+++ api/kogito-events-api/src/main/java/org/kie/kogito/event/KogitoThreadPoolFactory.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.event;
+
+import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.atomic.AtomicInteger;
+
+public class KogitoThreadPoolFactory implements ThreadFactory {
+    private final AtomicInteger counter = new AtomicInteger(1);
+    private String threadNamePrefix;
+
+    public KogitoThreadPoolFactory(String threadNamePrefix) {
+        this.threadNamePrefix = threadNamePrefix;
+    }
+
+    @Override
+    public Thread newThread(Runnable r) {
+        String threadName = threadNamePrefix + "-" + counter.getAndIncrement();
+        Thread th = new Thread(r);
+        th.setName(threadName);
+        th.setDaemon(true);
+        return th;
+    }
+}
