+++ quarkus/addons/messaging/common/src/main/java/org/kie/kogito/addon/quarkus/messaging/common/QuarkusEventExecutorServiceFactory.java
@@ -15,7 +15,6 @@
  */
 package org.kie.kogito.addon.quarkus.messaging.common;
 
-import java.util.Optional;
 import java.util.concurrent.ExecutorService;
 
 import javax.enterprise.context.ApplicationScoped;
@@ -23,8 +22,7 @@ import javax.inject.Inject;
 
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.kie.kogito.event.EventExecutorServiceFactory;
-import org.kie.kogito.event.KogitoEmitterController;
-import org.kie.kogito.event.KogitoEventExecutor;
+import org.kie.kogito.event.KogitoEventStreams;
 
 import io.quarkus.arc.DefaultBean;
 
@@ -32,16 +30,17 @@ import io.quarkus.arc.DefaultBean;
 @DefaultBean
 public class QuarkusEventExecutorServiceFactory implements EventExecutorServiceFactory {
 
-    @ConfigProperty(name = KogitoEventExecutor.MAX_THREADS_PROPERTY, defaultValue = KogitoEventExecutor.DEFAULT_MAX_THREADS)
+    @ConfigProperty(name = KogitoEventStreams.MAX_THREADS_PROPERTY, defaultValue = KogitoEventStreams.DEFAULT_MAX_THREADS)
     int numThreads;
 
-    @ConfigProperty(name = KogitoEventExecutor.QUEUE_SIZE_PROPERTY, defaultValue = KogitoEventExecutor.DEFAULT_QUEUE_SIZE)
+    @ConfigProperty(name = KogitoEventStreams.QUEUE_SIZE_PROPERTY, defaultValue = KogitoEventStreams.DEFAULT_QUEUE_SIZE)
     int queueSize;
 
     @Inject
-    KogitoEmitterController emitterStatus;
+    QuarkusEmitterController emitterStatus;
 
+    @Override
     public ExecutorService getExecutorService(String channelName) {
-        return KogitoEventExecutor.getEventExecutor(numThreads, queueSize, Optional.of(emitterStatus), channelName);
+        return new QuarkusEventThreadPool(numThreads, queueSize, emitterStatus, channelName);
     }
 }
