+++ springboot/addons/messaging/implementation/src/main/java/org/kie/kogito/addon/cloudevents/spring/SpringEventExecutorServiceFactory.java
@@ -15,23 +15,29 @@
  */
 package org.kie.kogito.addon.cloudevents.spring;
 
+import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.ExecutorService;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
 import org.kie.kogito.event.EventExecutorServiceFactory;
-import org.kie.kogito.event.KogitoEventExecutor;
+import org.kie.kogito.event.KogitoEventStreams;
+import org.kie.kogito.event.KogitoThreadPoolFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class SpringEventExecutorServiceFactory implements EventExecutorServiceFactory {
 
-    @Value("${" + KogitoEventExecutor.MAX_THREADS_PROPERTY + ":#{" + KogitoEventExecutor.DEFAULT_MAX_THREADS + "}}")
+    @Value("${" + KogitoEventStreams.MAX_THREADS_PROPERTY + ":#{" + KogitoEventStreams.DEFAULT_MAX_THREADS + "}}")
     int numThreads;
 
-    @Value("${" + KogitoEventExecutor.QUEUE_SIZE_PROPERTY + ":#{" + KogitoEventExecutor.DEFAULT_QUEUE_SIZE + "}}")
+    @Value("${" + KogitoEventStreams.QUEUE_SIZE_PROPERTY + ":#{" + KogitoEventStreams.DEFAULT_QUEUE_SIZE + "}}")
     int queueSize;
 
+    @Override
     public ExecutorService getExecutorService(String channelName) {
-        return KogitoEventExecutor.getEventExecutor(numThreads, queueSize);
+        return new ThreadPoolExecutor(1, numThreads, 1L, TimeUnit.MINUTES, new ArrayBlockingQueue<>(queueSize), new KogitoThreadPoolFactory(KogitoEventStreams.THREAD_NAME),
+                new ThreadPoolExecutor.CallerRunsPolicy());
     }
 }
